vision api/raspberry_pi/requirements.txt
opencv-python>=4.8.0
google-cloud-vision>=3.4.0
pyserial>=3.5
Pillow>=10.0.0

